test(config): add missing error path tests for GPU and GGUF detection
- Consolidated tests from tests/test_config_coverage.py into tests/test_config.py
- Added comprehensive tests for _detect_gpu (success, failure, import error, runtime exception)
- Added tests for _has_gguf_support (success, import error)
- Fixed coverage for various rerank resolution branches by clearing env vars in tests
- Achieved 99% coverage for src/mnemo_mcp/config.py
- Removed redundant tests/test_config_coverage.py file
